refactor: Move path expression functions into path_exp module
Specifically: `matches_token`, `calc_path_weight`, and `path_length`.
