+++ rust/pact_models/src/matchingrules.rs
@@ -17,7 +17,7 @@ use serde_json::{json, Value};
 use crate::{HttpStatus, PactSpecification};
 use crate::generators::{Generator, GeneratorCategory, Generators};
 use crate::json_utils::{json_to_num, json_to_string};
-use crate::path_exp::{parse_path_exp, PathToken};
+use crate::path_exp::{calc_path_weight, path_length};
 
 /// Set of all matching rules
 #[derive(Serialize, Deserialize, Debug, Clone, Eq)]
@@ -520,60 +520,6 @@ fn hash_and_partial_eq_for_matching_rule() {
   expect!(&ac7).to_not(be_equal_to(&ac6));
   expect!(&ac7).to_not(be_equal_to(&ac1));
 }
-
-fn matches_token(path_fragment: &str, path_token: &PathToken) -> usize {
-  match path_token {
-    PathToken::Root if path_fragment == "$" => 2,
-    PathToken::Field(name) if path_fragment == name => 2,
-    PathToken::Index(index) => match path_fragment.parse::<usize>() {
-      Ok(i) if *index == i => 2,
-      _ => 0
-    },
-    PathToken::StarIndex => match path_fragment.parse::<usize>() {
-      Ok(_) => 1,
-      _ => 0
-    },
-    PathToken::Star => 1,
-    _ => 0
-  }
-}
-
-/// Calculates the path weight for a path expression and a given path. Returns a tuple of the
-/// calculated weight and the number of path tokens matched
-pub fn calc_path_weight(path_exp: &str, path: &[&str]) -> (usize, usize) {
-  let weight = match parse_path_exp(path_exp) {
-    Ok(path_tokens) => {
-      trace!("Calculating weight for path tokens '{:?}' and path '{:?}'", path_tokens, path);
-      if path.len() >= path_tokens.len() {
-        (
-          path_tokens.iter().zip(path.iter())
-            .fold(1, |acc, (token, fragment)| acc * matches_token(fragment, token)),
-          path_tokens.len()
-        )
-      } else {
-        (0, path_tokens.len())
-      }
-    },
-    Err(err) => {
-      warn!("Failed to parse path expression - {}", err);
-      (0, 0)
-    }
-  };
-  trace!("Calculated weight {:?} for path '{}' and '{:?}'", weight, path_exp, path);
-  weight
-}
-
-/// Parses the path expression and returns the number of tokens in the path
-pub fn path_length(path_exp: &str) -> usize {
-  match parse_path_exp(path_exp) {
-    Ok(path_tokens) => path_tokens.len(),
-    Err(err) => {
-      warn!("Failed to parse path expression - {}", err);
-      0
-    }
-  }
-}
-
 
 /// Enumeration to define how to combine rules
 #[derive(Serialize, Deserialize, PartialEq, Debug, Clone, Eq, Hash, PartialOrd, Ord)]
@@ -1777,81 +1723,6 @@ mod tests {
             }
         };
     expect!(matchers.wildcard_matcher_is_defined("body", &vec!["$", "a", "b", "c"])).to(be_false());
-  }
-
-  #[test]
-  fn matches_token_test_with_root() {
-    expect!(matches_token("$", &PathToken::Root)).to(be_equal_to(2));
-    expect!(matches_token("path", &PathToken::Root)).to(be_equal_to(0));
-    expect!(matches_token("*", &PathToken::Root)).to(be_equal_to(0));
-  }
-
-  #[test]
-  fn matches_token_test_with_field() {
-    expect!(matches_token("$", &PathToken::Field("path".to_string()))).to(be_equal_to(0));
-    expect!(matches_token("path", &PathToken::Field("path".to_string()))).to(be_equal_to(2));
-  }
-
-  #[test]
-  fn matches_token_test_with_index() {
-    expect!(matches_token("$", &PathToken::Index(2))).to(be_equal_to(0));
-    expect!(matches_token("path", &PathToken::Index(2))).to(be_equal_to(0));
-    expect!(matches_token("*", &PathToken::Index(2))).to(be_equal_to(0));
-    expect!(matches_token("1", &PathToken::Index(2))).to(be_equal_to(0));
-    expect!(matches_token("2", &PathToken::Index(2))).to(be_equal_to(2));
-  }
-
-  #[test]
-  fn matches_token_test_with_index_wildcard() {
-    expect!(matches_token("$", &PathToken::StarIndex)).to(be_equal_to(0));
-    expect!(matches_token("path", &PathToken::StarIndex)).to(be_equal_to(0));
-    expect!(matches_token("*", &PathToken::StarIndex)).to(be_equal_to(0));
-    expect!(matches_token("1", &PathToken::StarIndex)).to(be_equal_to(1));
-  }
-
-  #[test]
-  fn matches_token_test_with_wildcard() {
-    expect!(matches_token("$", &PathToken::Star)).to(be_equal_to(1));
-    expect!(matches_token("path", &PathToken::Star)).to(be_equal_to(1));
-    expect!(matches_token("*", &PathToken::Star)).to(be_equal_to(1));
-    expect!(matches_token("1", &PathToken::Star)).to(be_equal_to(1));
-  }
-
-  #[test]
-  fn matches_path_matches_root_path_element() {
-    expect!(calc_path_weight("$", &vec!["$"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$", &vec![]).0 > 0).to(be_false());
-  }
-
-  #[test]
-  fn matches_path_matches_field_name() {
-    expect!(calc_path_weight("$.name", &vec!["$", "name"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$['name']", &vec!["$", "name"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$.name.other", &vec!["$", "name", "other"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$['name'].other", &vec!["$", "name", "other"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$.name", &vec!["$", "other"]).0 > 0).to(be_false());
-    expect!(calc_path_weight("$.name", &vec!["$", "name", "other"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$.other", &vec!["$", "name", "other"]).0 > 0).to(be_false());
-    expect!(calc_path_weight("$.name.other", &vec!["$", "name"]).0 > 0).to(be_false());
-  }
-
-  #[test]
-  fn matches_path_matches_array_indices() {
-    expect!(calc_path_weight("$[0]", &vec!["$", "0"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$.name[1]", &vec!["$", "name", "1"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$.name", &vec!["$", "0"]).0 > 0).to(be_false());
-    expect!(calc_path_weight("$.name[1]", &vec!["$", "name", "0"]).0 > 0).to(be_false());
-    expect!(calc_path_weight("$[1].name", &vec!["$", "name", "1"]).0 > 0).to(be_false());
-  }
-
-  #[test]
-  fn matches_path_matches_with_wildcard() {
-    expect!(calc_path_weight("$[*]", &vec!["$", "0"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$.*", &vec!["$", "name"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$.*.name", &vec!["$", "some", "name"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$.name[*]", &vec!["$", "name", "0"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$.name[*].name", &vec!["$", "name", "1", "name"]).0 > 0).to(be_true());
-    expect!(calc_path_weight("$[*]", &vec!["$", "name"]).0 > 0).to(be_false());
   }
 
   #[test]
